refactor: extract Rabbitmq const into class
