+++ src/main/java/org/example/chat/config/rabbitmq/RabbitMQConstants.java
@@ -1,0 +1,9 @@
+package org.example.chat.config.rabbitmq;
+
+public class RabbitMQConstants {
+
+    public static final String CHAT_QUEUE_NAME = "chat.queue";
+    public static final String CHAT_EXCHANGE_NAME = "chat.exchange";
+    public static final String ROUTING_KEY = "room.*";
+
+}
