+++ src/main/java/org/example/chat/controller/StompRabbitController.java
@@ -2,9 +2,8 @@ package org.example.chat.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.example.chat.config.rabbitmq.RabbitMQConstants;
 import org.example.chat.dto.ChatDto;
-import org.example.chat.service.ChatService;
-import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.messaging.handler.annotation.DestinationVariable;
 import org.springframework.messaging.handler.annotation.MessageMapping;
@@ -18,8 +17,6 @@ import java.time.LocalDateTime;
 public class StompRabbitController {
 
     private final RabbitTemplate template;
-    private final static String CHAT_EXCHANGE_NAME = "chat.exchange";
-
 
     @MessageMapping("chat.enter.{chatRoomId}")
     public void enter(ChatDto chatDto, @DestinationVariable String chatRoomId) {
@@ -27,7 +24,7 @@ public class StompRabbitController {
         chatDto.setRegDate(LocalDateTime.now());
 
         // exchange
-        template.convertAndSend(CHAT_EXCHANGE_NAME, "room." + chatRoomId, chatDto);
+        template.convertAndSend(RabbitMQConstants.CHAT_EXCHANGE_NAME, "room." + chatRoomId, chatDto);
         // template.convertAndSend("room." + chatRoomId, chat); //queue
         // template.convertAndSend("amq.topic", "room." + chatRoomId, chat); //topic
     }
@@ -37,7 +34,7 @@ public class StompRabbitController {
     public void send(ChatDto chatDto, @DestinationVariable String chatRoomId) {
         chatDto.setRegDate(LocalDateTime.now());
 
-        template.convertAndSend(CHAT_EXCHANGE_NAME, "room." + chatRoomId, chatDto);
+        template.convertAndSend(RabbitMQConstants.CHAT_EXCHANGE_NAME, "room." + chatRoomId, chatDto);
         //template.convertAndSend( "room." + chatRoomId, chat);
         //template.convertAndSend("amq.topic", "room." + chatRoomId, chat);
     }
